Make cgo map deal in pointers, not ints
To round-trip values through the _private void* field in the
rd_kafka_message_t type, we currently generate an int in cgoPut(), save
the value in a map, and cast the int to a uintptr_t when passing into
cgo, which is then cast (inside C code) to a void* before sending to
librdkafka.

To add support for passing _private in the batch-produce case, we need
to be able to assign the _private field (of type void*/unsafe.Pointer)
in Go, rather than in C. In Go, it is illegal to cast any int to an
unsafe.Pointer (even if it is never dereferenced).

Therefore, we must change cgoPut/cgoGet to deal with unsafe.Pointer
tokens, rather than int tokens. The most straightforward way to generate
such pointers which are guaranteed to never point to Go objects is to
allocate them on the C heap with C.malloc.

To avoid a bunch of pointless malloc/free calls, we also implement a
token cache so that these pointers can be reused.

+++ kafka/event.go
@@ -177,7 +177,7 @@ func (h *handle) eventPoll(ctx context.Context) (Event, error) {
 
 			if rkmessage._private != nil {
 				// Find cgoif by id
-				cg, found := h.cgoGet((int)((uintptr)(rkmessage._private)))
+				cg, found := h.cgoGet(rkmessage._private)
 				if found {
 					cdr := cg.(cgoDr)
